+++ pkgdcl.lisp
@@ -82,15 +82,9 @@
    #:find-option                        ; finds an option, not a descriptor
 
    ;; Conditions
-   #:undefined-field-type
-   #:undefined-input-type
-   #:undefined-output-type
-   #:undefined-stream-type
-   #:undefined-type
-   #:unknown-enum-error
-   #:error-type-name
-   #:error-field
-   #:error-method
+   #:protobuf-error
+   #:unknown-type
+   #:unknown-field-type
 
    ;; Extensions
    #:get-extension